Added missing Type
